Fix incorrect state handling for disconnects
Under the previous code, we were observing multiple state changes in the
connection - not just the network path update, but also the backend.
Killing the connection on every error was overkill and led to bugs.

With this new code it's the hope that Stealth connections will be stable
and users in countries practicing censorship will be able to access
restricted information.

Go with changing `.requiresConnection` to not be a satisfiable path.
Home <-> cell network transitions were causing issues otherwise. In
Russia this should help with resilience on filtered networks.