Bug: Fix incorrect IP address allocation size
Documentation for `withMemoryRebound(to:capacity:_:)` is available [here]
(https://developer.apple.com/documentation/swift/unsafepointer/withmemoryrebound(to:capacity:_:))

According to it, the `capacity` parameter is specified as "the number of
instances of T in the re-bound region," and not the total size of the rebound
struct. This can lead to crashes in the extension with the following error:

`Fatal error: self must be a properly aligned pointer for types Pointee and T`

Since the subsequent line in the code only reads `sizeof(in_addr)` or
`sizeof(in6_addr)` anyway, change the `capacity` parameter to just be a count
of 1.